move share recording to it's own file
